feat: #76 single out numerics
