+++ hamilflow/models/kepler_problem/model.py
@@ -10,9 +10,10 @@ from ...maths.trigonometrics import acos_with_shift
 from .dynamics import (
     tau_of_u_elliptic,
     tau_of_u_hyperbolic,
-    tau_of_u_parabolic,
-    u_of_tau,
+    tau_of_u_parabolic
+    
 )
+from .numerics import u_of_tau
 
 if TYPE_CHECKING:
     from typing import Collection, Mapping
